Crate noir_wasm

Modules

compile 🔒

compile_new 🔒

errors 🔒

Structs

BuildInfo

CompilerContext

This is a wrapper class that is wasm-bindgen compatible We do not use js_name and rename it like CrateId because then the impl block is not picked up in javascript.

CrateIDWrapper

Constants

BUILD_INFO 🔒

Functions

build_info

compile_contract

compile_contract_

This is a method that exposes the same API as compile But uses the Context based APi internally

compile_program

compile_program_

This is a method that exposes the same API as compile But uses the Context based APi internally

init_log_level