noir_wasm
Module compile

Structs

DependencyGraph 🔒

JsCompileContractResult

JsCompileProgramResult

JsDependencyGraph

PathToFileSourceMap

This map contains the paths of all of the files in the entry-point crate and the transitive dependencies of the entry-point crate.

Functions

add_noir_lib 🔒

compile_contract

compile_program

file_manager_with_source_map 🔒

prepare_context 🔒

process_dependency_graph 🔒